Add multiline log matching
This is related to #628 and #608 and should fix both issues.

Also, some refactoring has been performed for the wait checkers which were extracted into an extra package.
