+++ doc/changelog.md
@@ -8,7 +8,8 @@
   - When using properties for configuration only build when `from` or `fromExt` is set (#736)
   - Add new mojo "docker:save" for saving the image to a file (#687)
   - Check whether a temporary tag could be removed and throw an error if not (#725)
-  
+  - Allow multi line matches in log output (#628)
+    
 * **0.20.0** (2017-02-17)
   - Removed `build-nofork` and `source-nofork` in favor for a more direct solution which prevents forking of the lifecycle. Please refer the documentation, chapter "Assembly" for more information about this.
