+++ src/main/java/io/fabric8/maven/docker/wait/WaitChecker.java
@@ -1,0 +1,11 @@
+package io.fabric8.maven.docker.wait;
+
+/**
+ * @author roland
+ * @since 25/03/2017
+ */
+public interface WaitChecker {
+    boolean check();
+
+    void cleanUp();
+}
