+++ src/test/java/io/fabric8/maven/docker/service/RunServiceTest.java
@@ -6,7 +6,7 @@ import java.util.*;
 
 import io.fabric8.maven.docker.log.LogOutputSpec;
 import io.fabric8.maven.docker.util.Logger;
-import io.fabric8.maven.docker.util.WaitUtil;
+import io.fabric8.maven.docker.wait.WaitUtil;
 import io.fabric8.maven.docker.access.*;
 import io.fabric8.maven.docker.config.*;
 import io.fabric8.maven.docker.log.LogOutputSpecFactory;
