+++ src/test/java/io/fabric8/maven/docker/util/WaitUtilTest.java
@@ -9,10 +9,10 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.TimeoutException;
 
 import com.sun.net.httpserver.*;
+import io.fabric8.maven.docker.wait.*;
 import org.junit.*;
 
 import io.fabric8.maven.docker.config.WaitConfiguration;
-import io.fabric8.maven.docker.util.WaitUtil.WaitTimeoutException;
 
 import static org.junit.Assert.*;
 
@@ -30,13 +30,13 @@ public class WaitUtilTest {
 
     @Test(expected = TimeoutException.class)
     public void httpFail() throws TimeoutException {
-        WaitUtil.HttpPingChecker checker = new WaitUtil.HttpPingChecker("http://127.0.0.1:" + port + "/fake-context/");
+        HttpPingChecker checker = new HttpPingChecker("http://127.0.0.1:" + port + "/fake-context/");
         long waited = WaitUtil.wait(500, checker);
     }
 
     @Test
     public void httpSuccess() throws TimeoutException {
-        WaitUtil.HttpPingChecker checker = new WaitUtil.HttpPingChecker(httpPingUrl);
+        HttpPingChecker checker = new HttpPingChecker(httpPingUrl);
         long waited = WaitUtil.wait(700, checker);
         assertTrue("Waited less than 700ms: " + waited, waited < 700);
     }
@@ -44,21 +44,21 @@ public class WaitUtilTest {
     @Test
     public void httpSuccessWithStatus() throws TimeoutException {
         for (String status : new String[] { "200", "200 ... 300", "200..250" }) {
-            long waited = WaitUtil.wait(700, new WaitUtil.HttpPingChecker(httpPingUrl, WaitConfiguration.DEFAULT_HTTP_METHOD, status));
+            long waited = WaitUtil.wait(700, new HttpPingChecker(httpPingUrl, WaitConfiguration.DEFAULT_HTTP_METHOD, status));
             assertTrue("Waited less than  700ms: " + waited, waited < 700);
         }
     }
 
     @Test(expected = TimeoutException.class)
     public void httpFailWithStatus() throws TimeoutException {
-        WaitUtil.wait(700, new WaitUtil.HttpPingChecker(httpPingUrl, WaitConfiguration.DEFAULT_HTTP_METHOD, "500"));
+        WaitUtil.wait(700, new HttpPingChecker(httpPingUrl, WaitConfiguration.DEFAULT_HTTP_METHOD, "500"));
     }
 
     @Test
     public void httpSuccessWithGetMethod() throws Exception {
         serverMethodToAssert = "GET";
         try {
-            WaitUtil.HttpPingChecker checker = new WaitUtil.HttpPingChecker(httpPingUrl, "GET", WaitConfiguration.DEFAULT_STATUS_RANGE);
+            HttpPingChecker checker = new HttpPingChecker(httpPingUrl, "GET", WaitConfiguration.DEFAULT_STATUS_RANGE);
             long waited = WaitUtil.wait(700, checker);
             assertTrue("Waited less than 500ms: " + waited, waited < 700);
         } finally {
@@ -68,13 +68,13 @@ public class WaitUtilTest {
 
     @Test
     public void tcpSuccess() throws TimeoutException {
-        WaitUtil.TcpPortChecker checker = new WaitUtil.TcpPortChecker("localhost", Collections.singletonList(port));
+        TcpPortChecker checker = new TcpPortChecker("localhost", Collections.singletonList(port));
         long waited = WaitUtil.wait(700, checker);
         assertTrue("Waited less than 700ms: " + waited, waited < 700);
     }
 
     @Test
-    public void cleanupShouldBeCalledAfterMatchedExceptation() throws WaitUtil.WaitTimeoutException {
+    public void cleanupShouldBeCalledAfterMatchedExceptation() throws WaitTimeoutException {
         StubWaitChecker checker = new StubWaitChecker(true);
         WaitUtil.wait(0, checker);
         assertTrue(checker.isCleaned());
@@ -86,7 +86,7 @@ public class WaitUtilTest {
         try {
             WaitUtil.wait(0, checker);
             fail("Failed expectation expected");
-        } catch (WaitUtil.WaitTimeoutException e) {
+        } catch (WaitTimeoutException e) {
             assertTrue(checker.isCleaned());
         }
     }
@@ -115,7 +115,7 @@ public class WaitUtilTest {
         });
     }
 
-    private static class StubWaitChecker implements WaitUtil.WaitChecker {
+    private static class StubWaitChecker implements WaitChecker {
 
         private final boolean checkResult;
         private boolean cleaned = false;
@@ -164,7 +164,7 @@ public class WaitUtilTest {
 
         // preload - first time use almost always lasts much longer (i'm assuming its http client initialization behavior)
         try {
-            WaitUtil.wait(700, new WaitUtil.HttpPingChecker(httpPingUrl));
+            WaitUtil.wait(700, new HttpPingChecker(httpPingUrl));
         } catch (TimeoutException exp) {
 
         }
